+++ src/part/mod.rs
@@ -2,7 +2,7 @@ pub mod archive;
 pub mod elf;
 pub mod fhs;
 pub mod folio;
-pub mod lint;
+pub mod soname;
 pub mod prune;
 pub mod store;
 pub mod version;